Add the schema migration/installation changes.
This doesn't fix a direct upgrade path but devs can use the CLI tool
with a command of ./cli.php schema:updatesql to see the SQL changes,
or ./cli.php schema:update to run it directly. There is also an untested
route for running this in the browser via GET params but this is less
than ideal.